// src/lib/types/openclaw.ts
export interface WorkspaceFile {
    name: string
    path: string
    date: string
    agentId: string
}